refactor: build-previews.mts を .mjs へ変換し TS 型構文を除去
tsconfig・@types/node 追加なしでエディタ型エラーを解消するため、
他のスクリプトと揃えて素の JS（.mjs）に統一した。
package.json の build:previews / build:previews:landing 参照も合わせて更新。
