+++ package.json
@@ -29,7 +29,8 @@
     "build:templates": "turbo run build --filter=./templates/*/* --filter=./templates/*/*/*",
     "build:template": "sh -c 'turbo run build --filter=\"$1\"' --",
     "build:template:en": "sh -c 'node scripts/build-template-lang.mjs \"$1\" en' --",
-    "build:previews": "pnpm build:core && pnpm build:ui && tsx scripts/build-previews.mts",
+    "build:previews": "pnpm build:core && pnpm build:ui && tsx scripts/build-previews.mjs",
+    "build:previews:landing": "tsx scripts/build-previews.mjs --landing-only",
     "preview:template": "sh -c 'pnpm --filter \"$1\" preview' --",
     "deploy:template": "sh -c 'pnpm build:template \"$1\" && dir=$(pnpm --filter \"$1\" exec pwd) && node scripts/write-noindex-headers.mjs \"$dir/dist\" && wrangler pages deploy \"$dir/dist\" --project-name=\"lism-$1\" --branch=main --commit-dirty=true --commit-message=\"deploy $1\"' --",
     "deploy:previews": "pnpm build:previews && wrangler pages deploy .preview/merged --project-name=lism-templates --branch=main --commit-dirty=true --commit-message=\"deploy previews\"",
